Initial commit. Will parse code from a file.
Uses the JComp library and the WindowLib api to create a window, draw a
text area, and render formatted code onto the textarea, reading from a
built-in file. This is by no means functional as an IDE, but it is a
good proof-of-concept and this is a good break point considering the
massive refactoring I now need to do with WindowLib.